created another timer test
